Usbmaker
Created a tool that will list available drives and ask user to select drive to install macOS recovery files. This downloads Opencore and uses the macrecovery.py script to download the MacOS files.

Run USBMAKER.bat as administrator to run the file.

WARNING! There are no safety checks in place so it is entirely possible to delete any drives detected if you select the wrong drive. This will not happen immediately, but since this is proof-of-concept, much more checks will be added later on.